# add an async loop to speed up the pollers