--- a/internal/services/postgres/postgresql_flexible_server_firewall_rules_resource.go
+++ b/internal/services/postgres/postgresql_flexible_server_firewall_rules_resource.go
@@ -8,6 +8,7 @@ import (
 	"fmt"
 	"log"
 	"strings"
+	"sync"
 	"time"
 
 	"github.com/hashicorp/go-azure-helpers/lang/response"
@@ -190,11 +191,19 @@ func resourcePostgresqlFlexibleServerFirewallRulesCreateUpdate(d *pluginsdk.Reso
 		pollers = append(pollers, poller.Poller)
 	}
 
+	wg := sync.WaitGroup{}
+
 	for _, poller := range pollers {
-		if err := poller.PollUntilDone(ctx); err != nil {
-			return fmt.Errorf("polling after CreateOrUpdate: %+v", err)
-		}
+		wg.Add(1)
+		go func() {
+			defer wg.Done()
+			if err := poller.PollUntilDone(ctx); err != nil {
+				fmt.Errorf("polling after CreateOrUpdate: %+v", err)
+			}
+		}()
+
 	}
+	wg.Wait()
 
 	d.SetId(id.ID())
 	return resourcePostgresqlFlexibleServerFirewallRulesRead(d, meta)