--- a/internal/services/postgres/postgresql_flexible_server_firewall_rules_resource.go
+++ b/internal/services/postgres/postgresql_flexible_server_firewall_rules_resource.go
@@ -288,12 +288,19 @@ func resourcePostgresqlFlexibleServerFirewallRulesDelete(d *pluginsdk.ResourceDa
 		pollers = append(pollers, poller.Poller)
 	}
 
+	wg := sync.WaitGroup{}
+
 	for _, poller := range pollers {
-		if err := poller.PollUntilDone(ctx); err != nil {
-			return fmt.Errorf("polling after Delete: %+v", err)
-		}
+		wg.Add(1)
+		go func() {
+			defer wg.Done()
+			if err := poller.PollUntilDone(ctx); err != nil {
+				fmt.Errorf("polling after Delete: %+v", err)
+			}
+		}()
 	}
 
+	wg.Wait()
 	return nil
 }