Testing out some changes to make the polling a bit faster.

With a config adding three firewall rules like so:
```
locals {
  firewall_rules = [
{
    name = "test-rule-1"
    end_ip_address = "x.x.x.x"
    start_ip_address = "x.x.x.x"
  },{
        name = "test-rule-2"
    end_ip_address = "x.x.x.y"
    start_ip_address = "x.x.x.y"
  },{
        name = "test-rule-3"
    end_ip_address = "x.x.x.z"
    start_ip_address = "x.x.x.z"
  }
  ]
}

resource "azurerm_postgresql_flexible_server_firewall_rules" "test-fw" {
  server_id = data.azurerm_postgresql_flexible_server.example.id
  dynamic "firewall_rule" {
    for_each = local.firewall_rules
    content {
      name = firewall_rule.value.name
      start_ip_address = firewall_rule.value.start_ip_address
      end_ip_address = firewall_rule.value.end_ip_address
    }
  }

}
```

Without this change, adding them took three minutes:

```
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Creating...
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Still creating... [10s elapsed]
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Still creating... [20s elapsed]
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Still creating... [30s elapsed]
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Still creating... [40s elapsed]
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Still creating... [50s elapsed]
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Still creating... [1m0s elapsed]
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Still creating... [1m10s elapsed]
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Still creating... [1m20s elapsed]
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Still creating... [1m30s elapsed]
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Still creating... [1m40s elapsed]
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Still creating... [1m50s elapsed]
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Still creating... [2m0s elapsed]
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Still creating... [2m10s elapsed]
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Still creating... [2m20s elapsed]
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Still creating... [2m30s elapsed]
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Still creating... [2m40s elapsed]
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Still creating... [2m50s elapsed]
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Still creating... [3m0s elapsed]
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Creation complete after 3m5s ]
```

Adding this change in gets that same config down to one minute:
```
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Creating...
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Still creating... [10s elapsed]
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Still creating... [20s elapsed]
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Still creating... [30s elapsed]
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Still creating... [40s elapsed]
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Still creating... [50s elapsed]
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Still creating... [1m0s elapsed]
azurerm_postgresql_flexible_server_firewall_rules.test-fw: Creation complete after 1m3s
```

Again, there might be a better way to do this, but it does seem positive. Let me know that you think!